+++ Module/changelog.md
@@ -1,3 +1,15 @@
 # Changelog Material Keys Module
+### v1.0.1 - 27-10-2020
+Fixes:
+<ul>
+<li>Fixed issue where LED would not be set properly when a playlist was playing</li>
+</ul>
+Additions:
+<ul>
+<li>'Playlist control' and 'playlist volume control' can now control up to 32 tracks per playlist</li>
+<li>Added method to see what page the user is on when using 'Playlist control', 'playlist volume control' and 'token health tracker'</li>
+<li>Added support for new FXmaster weather effects ('Snowstorm' and 'Rain without splash')</li>
+</ul>
+
 ### v1.0.0 - 21-10-2020
 Initial release